requirements.txt
pandas==2.2.2
pyyaml==6.0.2
typer==0.12.3
rich==13.7.1
pytest==8.2.0